+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Opening preferences from Extras menu open addon expandded
 ### Fixed
 - Error report messages
+- Auto add & remove prefix operator > wouldnt add and remove prefix properly
 
 ## [0.1.5] - 2022-09-13
 ### Added